refactor: rename content containers
