+++ src/lib/Scenes/SavedSearchAlert/containers/CreateSavedSearchContentContainerV1.tsx
@@ -1,34 +1,34 @@
 import { StackNavigationProp } from "@react-navigation/stack"
-import { ContentRefetchContainer_me } from "__generated__/ContentRefetchContainer_me.graphql"
+import { CreateSavedSearchContentContainerV1_me } from "__generated__/CreateSavedSearchContentContainerV1_me.graphql"
 import React from "react"
 import { createRefetchContainer, graphql, RelayRefetchProp } from "react-relay"
-import { Content } from "../Components/Content"
+import { CreateSavedSearchContent } from "../Components/CreateSavedSearchContent"
 import { CreateSavedSearchAlertNavigationStack, CreateSavedSearchAlertParams } from "../SavedSearchAlertModel"
 
 interface ContainerProps extends Omit<CreateSavedSearchAlertParams, "me" | "onClosePress"> {
   navigation: StackNavigationProp<CreateSavedSearchAlertNavigationStack, "CreateSavedSearchAlert">
   relay: RelayRefetchProp
-  me?: ContentRefetchContainer_me | null
+  me?: CreateSavedSearchContentContainerV1_me | null
 }
 
 const Container: React.FC<ContainerProps> = (props) => {
   const { me, ...other } = props
-  return <Content {...other} userAllowsEmails={me?.emailFrequency !== "none"} />
+  return <CreateSavedSearchContent {...other} userAllowsEmails={me?.emailFrequency !== "none"} />
 }
 
-export const ContentRefetchContainer = createRefetchContainer(
+export const CreateSavedSearchContentContainerV1 = createRefetchContainer(
   Container,
   {
     me: graphql`
-      fragment ContentRefetchContainer_me on Me {
+      fragment CreateSavedSearchContentContainerV1_me on Me {
         emailFrequency
       }
     `,
   },
   graphql`
-    query ContentRefetchContainerQuery {
+    query CreateSavedSearchContentContainerV1RefetchQuery {
       me {
-        ...ContentRefetchContainer_me
+        ...CreateSavedSearchContentContainerV1_me
       }
     }
   `
